requirements.txt
asgiref==3.8.1
attrs==23.2.0
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.85
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pathspec==0.10.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.0
termcolor==1.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
wcwidth==0.1.9
